More cast warnings

+++ Glaceon/VulkanRenderer/VulkanCommandPool.cpp
@@ -74,7 +74,8 @@ void VulkanCommandPool::RebuildCommandBuffers() {
 
   // destroy all old command buffers
   device.freeCommandBuffers(vk_command_pool_, 1, &vk_main_command_buffer_);
-  device.freeCommandBuffers(vk_command_pool_, vk_frame_command_buffers_.size(), vk_frame_command_buffers_.data());
+  device.freeCommandBuffers(vk_command_pool_, static_cast<uint32_t>(vk_frame_command_buffers_.size()),
+                            vk_frame_command_buffers_.data());
 
   vk_frame_command_buffers_.clear();
   std::vector<SwapChainFrame> swap_chain_frames = context_.GetVulkanSwapChain().GetSwapChainFrames();
